ability to use externally pre-computed features for classification
This commit adds a new button (Precomputed Features...) to the feature
selection applet. As an alternative to selecting features from the
pre-determined set of features using the feature selection dialog, you
can now specify a text file which contains a list of features you have
pre-computed and stored on disk.

After clicking on the "Precomputed Features" button, select a text file
with the following format:

  - one filename per line
  - each filename is an HDF5 file, with dimensions xyz, no channels
    dataset is called "data"

You can then use these features for learning and prediction.