Remove extra space, fixes #148, improves #6.
This commit changes the class of the central widget from QMainWindow
to QWidget. This needs to be reflected in centralWidget.ui file if
this has been copied.

In general, this commit reduces margins of layouts
(setContentsMargins 0,0,0,0) and re-layouts widgets such that the
GUI looks more homogenous and wastes less space.