Move "Add..." button to a row of the datasetDetail view.
In the data selection applet, the summary view includes the "Add..."
button on the last row of the table. This patch changes the rest of the
views to do the same.

Extra row logic used for the summary view is refactored into a new class
and used again for the datasetDetailedInfoTableModel.