DataSelection: Show add button in each row without data.
"Add..." button behaviour in the current data selection applet is to
fill the first empty lane for the given role with the specified file.
This can lead to confusion when displaying the add button on the last
row of the table.

This patch displays and "Add..." button on each empty row. The main
"Add..." button on the last row is preserved and renewed to "Add New..."