Modified run_each_unit_test script to not stop on first failure
currently, run_each_unit_test.sh will exit on the first failed test.
A code change might effect more tests. All those tests offer the
possibility to track down the bug/error, so testing should not stop
after encountering the first fail.

In this new variant, the script will continue to run each test,
accumulate the number of test runs that fail (NOT the number of
single test cases, but the number of nosetest executions that
return with with an errorcode != 0
At the end of the script, the failing tests are printed to stdout.