TrackingFeatureExtraction: do no overwrite inputslot value depending on data shape. Adresses #1432 in tracking workflows.
Overwriting the list of pre-selected features for tracking depending on the raw shape is too restrictive.
If the user first loads a 3D dataset but then, by adjusting the axestags, changes that to 2D+t, the 2D features should become available again.
If the feature slot value is overwritten but the original value is lost, this cannot be recovered. Hence here the initial value is
stored as member variable of the operator, and the input slot is overwritten with the correct value.