Do not format output filename on bad format string
`format_known_keys()` raises `ValueError` if the format string is malformed:
`{spam} {eggs` is malformed because the second field is not closed.
This can happen if a format string is coming from the user:
see https://github.com/ilastik/ilastik/issues/1761.

This commit allows non-strict behaviour of `format_known_keys()`, which is
disabled by default, and enables it in places that cause the issue above.