Always instantiate the CacheMemoryManager on start
Previously, the manager would only be instantiated on startup in
specific circumstances (when a refresh interval is explicitly set).
Therefore, for some headless workflows, the cache purging thread was
never started and the caches would grow uncontrollably.

Now it is always instantiated with the default interval (which can then
be changed later).