Closes #1770
Correctly serializes labels to be encoded as JSON. Also hides the
'Export' menu entry when not in --debug mode, instead of allowing
it to show up without any sub-menu entries.

It's worth noting, though, that the ex/importing of labels still
has some issues; exporting labels, clearing and then re-importing
may not reproduce the exact same labels in many cases.