added function to safely retrieve originalShape
Voluminas `slotMetaInfoDisplayWidget` would previously fail if `original_shape`
was not present in the meta dict. This is only the case if an
`OpReorderAxes` was somewhere added in the chain.
I added a function similar to `getOriginalAxisKeys()` for the shape:
`getOriginalShape()`.
Also added some comment to make it more obvious where those `original_` metadata
come from.