Adapts newAutocontext for uint16 output
The goal was to get the autocontext working with uint16 data.
Towards this I modified the opPixelClassfication to also produce uint16
probabilities as well as the newAutocontextWorkflow to also accept uint16
input and use the corresponding uint16 probabilities when the input is uint16.

This was tested by doing assembling and running an 2 stage autocontext sucessfully
as well as running the existing test suite.