Fixes 'Index Error' when adding new label to pixel 'Training' applet
The thresholding operator must update its GUI when new labels
are added to the pixel classification 'Training' applet. In order
to do that, the operator/applet was watching for changes in its
InputImage slot and it would try to update its widgets before its
InputChannelColors slot would be ready with the new label, causing
an "Index Error" when trying to find the new color of the new label.

This commit makes the applet watch instead for changes on the
InputChannelColors slot, so that it will always have the information
needed for updating the GUI when the watcher function fires.