Fixes PR observations and fixes multichannel pure-color mask issues
A user might want to use a (colored) atlas as a mask. If one of those colors
was pure (like red 255,0,0 ), then the previous code would think that predictions
should be masked for the 2nd and 3rd channel of the mask. This is probably not
what the user wanted, so now all channels of the mask are OR'ed together
to form the final mask which is then used to check for the need to compute
predicitons