Removed legacy WatershedSource, InvertWatershedSource
previously (pre ilastik 0.6???) it might have been possible (or planned) to
allow for selection of the source for the watershed (from raw, overlay,
filtered). The current implementation didn't allow for that choice, so all
references to that have been removed.
Serialization: if that stuff is present in older project, it is simply ignored.
This assumes that the filtered_raw data has always been the default source.