WIP: Returns exception in _applyTempOpSettingsRealOp as expected by accept()
There is so much (recursive) logic with trying to guide the user into fixing
their data/features that I'm starting to think we're doing something
fundamentally wrong. It seems to me that perhaps we should merge
"Data Selection" with 'Feature Selection", because those applets must
be compatible with each other and it makes little sense to have the user going
back and forth trying to fix either the data or the features if they don't match.

Also minor syntax fixes.