use time.perf_counter instead of os.times
previously the code used os.times()[4] (elapsed time since some point
in the past) to set start and stop times in order to report timing.
This doesn't work on windows (from the docs:

"See the Unix manual page times(2) or the corresponding Windows Platform
API documentation. On Windows, only user and system are known; the other
attributes are zero."

)

with time.perf_counter() this is portable.