Add StrictABC and use it for ShellABC (#2090)
`ShellABC` now uses `StrictABCMeta` metaclass and registers it's virtual
subclasses after their definitions.

PyQt5 uses a custom SIP metaclass for QObject and it's subclasses,
making it impossible to use plain `ABC` or `ABCMeta` for `ShellABC`
since `IlastikShell` inherits from `QMainWindow`.

It is possible to define a custom metaclass
`QObjectABCMeta(QObject.__class__, ABCMeta)`  and use it for `ShellABC`,
but in this case we always depend on `QObject`, even though some
subclasses should not use Qt at all (e.g. `HeadlessShell`).

We could use `ABC.register` to "virtually" subclass `ShellABC`. However,
because such subclass is not a real ABC subclass, it is possible to
create an instance even if some abstract methods are not implemented.

Therefore, we perform all necessary checks during the `register` call
in `StrictABCMeta`. This solution automatically provides proper
`__subclasshook__` and `__instancehook__`, together with more meaningful
error message.