Force FilesystemDatasetInfo when recovering from missing files
If the missing file was cause because the original link was relative
and the project file was moved, then the previous behavior was to
try to instantiate a RelativeFilesystemDatasetInfo from the fixed file
path. That file could bery well exist in a path that could not be
expressed relative to the current project file (i.e. in the same folder
or in any subfolder), raising an exception.

This patch forces fixed missing files to use FilesystemFatasetInfo,
which is an absolute path and therefore cannot cause this problem.

(cherry picked from commit 1608b8321e0d144d59b891f54167103ec6120b2c)