fix qt uic parsing; property -> getter/setter
The cause is PyQt5 accessing all class attributes via PyObject_GetAttr during initialization of the Gui Widget.
In our case, the class is not fully initialized by that time but class properties are still executed (via PyObject_GetAttr).

For some properties it is not possible to initialize properly.

So properties are made get/set functions in order to deal with it.

This error is also supposed to go away with qt 5.13 (not currently available via conda) (or pyside2).