Cleanup gui callbacks, workaround for batch-processing bug
This fixes batchprocessing in Gui mode for workflows with limits on the lane
numbers. Buttons in the summary view are deleted in batch processing to a bug
https://github.com/ilastik/ilastik/issues/2380. With this workaround, batch
processing at least works for all workflows again.

Also to be a bit more cautious callbacks that tough the Gui are cleaned up and
made sure to run in main thread.

fixes #2270