Make sure ilastik errors when trying to access missing data in headless
When running ilastik in headless, we do not require training data to be present.
However, ilastik will try to train its classifier, if it has not been saved to
the project file.
This can lead to training from all zeros when data is inaccessible.
This is of course not acceptable.