Fix edgetraining deserialization
deserialization carried the implicit assumption that always all lanes
contain values for the RAG. In cases where e.g. only a rag was saved for lane 2, deserialization would result in this Rag then being loaded to lane 0 :o.
This was then completely inconsistent - would result in IndexError in some cases.