configure logging only once
logging is configure whenever the main ilastik main function is executed
(`ilastik.app.main`). In tests this happens multiple times.

This change prevents multiple initialization of logging. For me this
reduced the log size per test run from 2.7MB to 270kB

a warning is issued if configuration is attempted a second time

Co-authored-by: Emil Melnikov <emilmelnikov@users.noreply.github.com>