fix multicut not updating result
changing the threshold value for the multicut when training the RF results in
desired outcomes.

Doing this, however, when not training the RF resulted in results not being updated.
Classifier cached would permanently be dirty - and invalidate the result.

Here, the final operator will ignore dirty propagation of the classifier if
Random forest is not trained.

Co-Authored-By: Tomaz Vieira <Tomaz-Vieira@users.noreply.github.com>