call compare function directly
this is the quickest way of going away from generator tests which are deprecated
by pytest.

Also sort of makes sense here, as different regions in the same filters are
compared. No need to have separate test cases for those.

For `test_features` it would be preferable to do this maybe with a fixture
as actually different filters are compared. For now keeping the direct call.

Number of tests for this file goes down from 33 to 8 (expected)