make model text edit non-editable on error/incompatible model
previously it was possible to edit content, delete everything - but it wouldn't
behave properly - keeping the old model source.

Now even on error you have to clear the model (this gives you a chance of
noting what model causes the problem and also the cause via the error message).

Also made sure live update is properly disabled on state change, and DL dialog
is modal.