Carving: gh2627 - Reinsert existing labels as new brushstrokes
This is a workaround because ilastiktools.setSeeds does not work as intended.

setSeeds currently writes 2 (i.e. "background") into the label array for both foreground and background labels. This bug was not a problem so far because the only usage in loadObject is immediately followed by the same reinsertion of the labels into the WriteSeeds input, which correctly overwrites the erroneous label array written by setSeeds.