Carving: Remove the alternative object readiness mechanism
Previously, HasSegmentation and dataIsStorable were used as two parallel mechanisms to ask the operator whether its current state can be saved as an object.

dataIsStorable was only tracking whether both types of labels/seeds exist, while HasSegmentation was tracking whether the object was segmented (which implied that both types of labels exist).

We make this explicit now and keep only one point to provide this information to outside opCarving.