move applet index saving to save method (#2660)
previously, closing the project would trigger update of currentApplet setting in project file. This had the unintentional side effect, that opening a file, doing something without saving, closing (without saving) would still alter the project file. This should not happen.

Now the currentApplet setting is only updated on the saveProject action (not on saveProjectAs, or saveProjectSnapshot).

Co-authored-by: Emil Melnikov <emilmelnikov@users.noreply.github.com>
Co-authored-by: Benedikt Best <63287233+btbest@users.noreply.github.com>