Refactor FilePathButton
Move complex update logic to a single function with an optional flag for avoiding computing `abbrevs`.

For a path with 6 abbreviations, the full update takes 0.5 ms, while skipping `abbrevs` takes only 0.05 ms.
It may not be a big difference, but event handlers for resize, paint and other such events should be as fast as possible.