functionality to ignore multiple dirtyprop from same source
Dirty propagation in ilastik goes depth-first. It is possible for dirtiness
to propagate to the same target via multiple upstream paths. This commit adds
the concept of a "modification time" (`mod_time` in the code) to encode sources
of dirtiness.
Operators will keep track of the last `mod_time` they've seen (during dirty
propagation) and can ignore it, if it is already known (see
`Operator.setAllDirty`).

In more complicated graphs like in autocontext, this solves heavy delays when
annotating.