functionality to ignore multiple dirty propagations from same source (#2694)
* functionality to ignore multiple dirtyprop from same source

Dirty propagation in ilastik goes depth-first. It is possible for dirtiness
to propagate to the same target via multiple upstream paths. This commit adds
the concept of a "modification time" (`mod_time` in the code) to encode sources
of dirtiness.
Operators will keep track of the last `mod_time` they've seen (during dirty
propagation) and can ignore it, if it is already known (see
`Operator.propagateDirtyIfNewModTime`).

In more complicated graphs like in autocontext, this solves heavy delays when
annotating.

  * all calls to `slot.setDirty` or `slot.setValue` without a mod_time will trigger
    a "new" mod_time.
  * `op._pending_dirty_mod_time` is only "active" during dirty prop
  * `op._previous_dirty_mod_time` holds the last completed dirtiness
    mod_time.
  * `op.propagateDirtyIfNewModTime`: note, this function will set all
     output slots dirty - for their entire extend (empty roi).