OpDataSelection: Clean up children instead of internal tracking
`self.children` has to be in the right order for this: The cleanup has to happen in reverse of the graph, as `cleanUp` throws if the op still has any outputs connected.
Not an issue now that we only expect one child from get_provider_slot, but would have been an issue a few commits ago.

See #2755