better "debug" for error in command line usage
it's sadly common to get not very helpful error messages when using
ilastik from the command line.
For me this often comes back to this error message complaining about
trailing file names.
It is absolutely not obvious what portion of the cmd is treated that way
so now we at least give a hint on what is interpreted incorrectly.

xref: https://github.com/ilastik/ilastik/issues/1830