Use exception chains with failing requests
Requests used to reraise the root cause exception, which made tracebacks
really hard to understand.
Now failures that occur in `Request`s will result in raising of
`RequestError`.
While this moves the root cause away from the last lines of the trace-
back, the flow/cause-chain is not much easier to follow with the user-
action being at the bottom of the traceback, and the root cause at the
top.
Following through this chain reflects graph connections.

Since we are Python3 only, also removed
`future.utils.raise_with_traceback` functionality.