fix tests for exception chaining
in previous implementation, exceptions where re-raised with the same
class as the root exception.
This way one could try-except `req.wait()` with the class of the root
exception.

With exception-chaining this is not the case anymore - but a similar
behavior can be achieved with the utility function `is_root_cause`,
to check an Exception class against the raised exception chain.