Ensure all object features get exported to table (#2814)
Ensure all object features get exported

Previously, only features computed in the 1st timeframe would get a column
in the export table. One would expect that with this bug there would be none ;).
Global features are always computed -> produce a table, but local ones, only if
there are objects.

This is more a bandaid then a proper fix. It would probably be better to
go from `selected` features (from the dialog) for column names, but currently
the plugin information is stripped from those :(

Co-authored-by: Benedikt Best <63287233+btbest@users.noreply.github.com>