OME-Zarr: Do not change output meta on existing store
When changing scale, OpDataSelection.setupOutputs instantiates a new OpInputDataReader and in turn a new instance of this op. Before that, for some reason, setupOutputs is called on the old instance, which runs into this if-branch. self.Scale is the old scale in this case, so setting e.g. shape here can bubble down the graph and put downstream slots (e.g. OpFeatureMatrixCache.LabelImage) into ready state too early, with metadata that can cause conflicts once the new instance is created and set up (e.g. assert on shape equality in OpFeatureMatrixCache.setupOutputs).