Make ModelSession to work only with single input and output tensors
There was the assumption that ModelSession should be able to handle multiple inputs and outputs, and each one of those with multiple shapes. This logic was left partially implemented, with a lot of assertions whenever ModelSession was used. At the same time, the interface of tiktorch server doesn't support multiple tensors and multiple shapes.

The feature of supporting this should be revised, and not be left partially implemented. To be able to continue with other features, ModelSession has been refactored to just work with the simple case of single tensors and shapes.