Default feature selection in object classification as in prior versions
with the inclusion of the sphericaltexture plugin, the
"All excl. location" button behavior changed - it would also include
those new features.
Here we change the behavior of the button to include only the default
features, as they were present in previous versions.

Timing-wise the sphericaltexture features are about a factor of 10x
slower to compute than our other features - so these should be selected
only if useful for the task (tbd by users).

* added dedicated `_setDefaults` function which will give the default
  behavior from previous versions
* "All excl. location" button renamed to "Default selection"
* removed any "location" logic from `_setAll`
* renamed `location` parameter to `exclude_location`, which is more
  expressive imo