OME-Zarr export: Account for subregion export
I.e. the user chose to export only a crop of the full output.
This means that if the input was multiscale, the export's shape cannot be used to infer which scale in the input it came from -> have to pass the active_scale through from the reader ops.

Additionally, if the crop does not start at 0 along any axis, CoordinateOffset will record the position relative to the full output. This corresponds to a translation-type coordinateTransformation in OME-Zarr.