oc: refactor BinaryAnything to SegmentationAnything
throughout oc and (and tracking) the code we were referring to the segmentation
input as binary, although both, label images and binary segmentations are
supported.

Also connected Binary Layer Display to Segmentationlayer display with
an optimized colormap

label image _has_ to be cached/accessed always, binary image not really.
Binary display can be achieved with a two color colormap.

As a consequence this should be friendlier for caching.

Also, display does not rely anymore on a binary image being present,
which could not be the case for label-type inputs.