oc: source Binary Image layer also from layer image
label image _has_ to be cached/accessed always, binary image not really.
Binary display can be achieved with a two color colormap.

As a consequence this should be friendlier for caching.

Also, display does not rely anymore on a binary image being present,
which could not be the case for label-type inputs.