OpRelabelConsecutive rewrite with cache
Rewritten OpRelabelConsecutive with a Cache compatible interface.
OpRelabelConsecutive5D caches the relabeled array and a mapping dict
in sync.

In particular, the methods related to data storage

* `_fetch_and_store_block`
* `_store_block_data`
* `setInSlot`
* `_execute_Output_impl`

as well as the memory management and reporting related methods

* `usedMemory`
* `freeBlock`

were copied and then adapted (so there is a bit of overlap).

Motivation is to support properly label images in object classification
where we have to relabel consecutively for vigra feature computation,
but also export the original labels in the table.