Make misconfigured OpMultichannelSelector.Output unready
There could be a mismatch, where SelectedChannels refer to channels not
present in the data. This raised an exception before. But this flow came
with the downside that the SelectedChannels input would ideally have to
be cleared during setupOutput calls which I would rather avoid.
Disconnecting the SelectedChannels slot before changes on Input can also
not be guaranteed.

This commit allows incompatible settings, but makes the output unready.
Ui can check this and correct.