Quickfix: Same object size limits in tracking and oc (#2958)
This commit sets the size filter initial values to the same ones than as in
thresholding. Also same initial value in all tracking flavors.

Tracking reuses some elements from Object Classification. E.g. the thresholding
applet. For the actual tracking there is another size-filter step, as there is
one in thresholding, too. These don't necessarily need to be in sync, but it's
better to start off that way.

fixes #2810