Use "show_in_startup_menu" attribute for workflow listing
This commit simplifies workflow listing.
Previously:
1. Workflows listed in `ilastik.workflows.AVAILABLE_WORKFLOWS`. Those
   were added to the start menu (irrespective of their `auto_register`).
2. Other workflows would be also added if their `auto_register` is `True`
3. The workflows named in
   `ilastik.shell.ilastikShell.IlastikShell_populateWorkflows` would be
   added first. Then the other ones.
At that stage it didn't make any difference if a workflow was listed in
`AVAILABLE_WORKFLOWS`, or the `auto_register` was `True`

`ilastik.workflows.AVAILABLE_WORKFLOWS` is removed -> it added an
additional priority layer that is not really necessary, since we do the
explicit listing in `ilastik.shell.ilastikShell.IlastikShell`.

The `auto_register` is renamed to `show_in_startup_menu` and now gets
the intended meaning - if set to `True` a workflow will be listed if the
workflow class was imported in `ilastik.workflows.__init__`. All
workflows imported there are available though for e.g. loading project
files, even if the workflow is not listed.

Our workflow listing is now simplified:

All workflows that are imported and have `show_in_startup_menu == True`
will show up in the menu. Whether it's in a specific group, or other
depends on `ilastik.shell.ilastikShell.IlastikShell_populateWorkflows`.