Add `DatasetOut` slot to OpDataSelection
This slot exposed the same information as the input `Dataset`, but it is
only connected if there are no data constraints violated.

Ui then can rely on this data being valid when connected.

DatasetDetailedInfoTable now listens to this slots, instead of `Dataset`
so no update gets triggered if data is not valid.

In the end, all consumers of OpDataSelection should call setValue only
on the Input `Dataset`, and retrieve values / listen for changes on
`DatasetOut` only.

Fixes #2967
Fixes #992