OpResize: Implement upscaling
Requires ensuring that the halo is at least enough for interpolation. Wasn't an issue for downscaling because antialiasing halo was always enough.

Seems the halo for order-1 needs to be at least one full pixel. A fractional "halo" (from ceil-rounding a reverse-scaled roi that had fractional starts or stops) still caused boundary artifacts.