Add `DatasetOut` slot to OpDataSelection (#2968)
This slot exposed the same information as the input `Dataset`, but it is
only connected if there are no data constraints violated. The slot is
intended to be connected to UI elements of the data selection table.
Ui then can rely on this data being valid when connected.

DatasetDetailedInfoTable now listens to this slots, instead of `Dataset`
so no update gets triggered if data is not valid.

Fixes #2967
Fixes #1956
Fixes #992