Lazyflow.Slot: Fix multi-inputs not signalling unreadiness (#3057)
This lets InputSlots with level>0 correctly determine their `oldReady`, so that `_sig_unready` is emitted at the end of `disconnect`. This is necessary because operators attach their `handleInputBecameUnready` directly to the InputSlot and not its `_subSlots`, so when `disconnect`ing the InputSlot, the operator's callback was never triggered.

(cherry picked from commit 185ff4b1bfaf7fdd0b8128d1dcda68eb309b272e)