Refactor: move setting widgets to HorizontalMainSplitter
This hides viewerControlStack and centralStack, mainControls, and
imageSelectionGroup from the shell. Methods have been added to
HorizontalMainSplitter to allow for the required interactions.

Chose not to make self.imageSelectionCombo private - it's used heavily,
also in tests.

Note on the unique object names: The "recording" feature doesn't exist
anymore. So these unique object names are not needed for that anymore.
I kept them as it often helps in debugging.

There is a slight change in behavior (no observable difference for me):
The centralWidget was replaced when the image index changed and inserted
in the position of the applet index, whereas viewerControls were all
stacked in the stack widget. I unified this, keeping both in their
respective StackWidgets. The references to the centralWidgets were being
kept alive anyway.

A (positive) side effect of this refactor: The "QLayout::removeWidget:
Cannot remove a null widget" message when closing projects is a thing of
the past.